review fix - rename some more testSuits and the contextLoader
