+++ README.MD
@@ -350,13 +350,13 @@ Same as it was at the bq view tests.
             * _{id from context.json}_.sql      
     * test_case2
     * jarvisTestSuite.json
-3. one or more .sql file somewhere in the project (testSuit.json queryPath) to make Bq table modification(s).
+3. one or more .sql file somewhere in the project (jarvisTestSuit.json queryPath) to make Bq table modification(s).
     
 Load to the BQ table and assert data are the same as it was at the bq view test.
 
 
 
-In BqQuery test the testSuit.json contains one or more BqQuery execution. contextType is always BqQuery, queryPath is the path feom the project folder to the .sql file.
+In BqQuery test the jarvisTestSuit.json contains one or more BqQuery execution. contextType is always BqQuery, queryPath is the path feom the project folder to the .sql file.
 
 jarvisTestSuite.json example
 ```
@@ -378,7 +378,7 @@ context.json example
 
 contexts.json (= CONTEXT_PATH)
 
-Make contexts.json file in test/resources to define parameters for the tests. LocalContext is for read local files, and the BigQuery context is for make the bigQuery job. The repositoryPath is the absolute path on the project folder. The test get the .sql file absolute path from the repositoryPath + testSuit.json queryPath.
+Make contexts.json file in test/resources to define parameters for the tests. LocalContext is for read local files, and the BigQuery context is for make the bigQuery job. The repositoryPath is the absolute path on the project folder. The test get the .sql file absolute path from the repositoryPath + jarvisTestSuit.json queryPath.
 ```
 [
   {
@@ -402,10 +402,10 @@ Make contexts.json file in test/resources to define parameters for the tests. Lo
 
 1. Set up the folder system with the files, pay attention of naming the folders and files, the json format and data in the json files.
 2. In IntegrationTestRunner.class set the CONTEXTS_PATH and the CONFIG_PATH. CONTEXTS_PATH is the absolute path of the contexts.json. CONFIG_PATH is the absolute path of the folder which has the test cases.
-3. Set up the contexts.json with LocalContext and add execution to the testSuit.json. LocalContext is for read local files.
+3. Set up the contexts.json with LocalContext and add execution to the jarvisTestSuit.json. LocalContext is for read local files.
 4. Run IntegrationTestRunner.class
 
-#### testSuits.json
+#### jarvisTestSuits.json
 
 1. caseAutoDetect
     * boolean
