+++ src/main/java/ai/aliz/jarvis/config/ExecutionActionConfig.java
@@ -1,4 +1,4 @@
-package ai.aliz.jarvis.jarvisconfig;
+package ai.aliz.jarvis.config;
 
 import lombok.Data;
 
@@ -6,13 +6,13 @@ import java.util.HashMap;
 import java.util.Map;
 
 @Data
-public class InitActionConfig implements StepConfig {
-    
-    private String system;
-    private InitActionType type;
-    
+public class ExecutionActionConfig implements StepConfig {
+
+
+    private String executionContext;
+    private ExecutionType type;
+
     private String descriptorFolder;
-    
+
     private final Map<String, Object> properties = new HashMap<>();
-    
 }
