+++ src/main/java/ai/aliz/jarvis/config/InitActionConfigFactory.java
@@ -1,4 +1,4 @@
-package ai.aliz.jarvis.jarvisconfig;
+package ai.aliz.jarvis.config;
 
 import lombok.Lombok;
