+++ src/main/java/ai/aliz/jarvis/service/JarvisRunnerService.java
@@ -5,19 +5,19 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import ai.aliz.jarvis.jarvisconfig.JarvisTestSuite;
+import ai.aliz.jarvis.config.JarvisTestSuite;
 import ai.aliz.jarvis.service.assertaction.AssertActionService;
 import ai.aliz.jarvis.service.executeaction.ExecuteActionService;
 import ai.aliz.jarvis.service.initaction.InitActionService;
-import ai.aliz.jarvis.jarvisconfig.JarvisTestCase;
-import ai.aliz.jarvis.jarvisconfig.JarvisConfigLoader;
+import ai.aliz.jarvis.config.JarvisTestCase;
+import ai.aliz.jarvis.config.ConfigLoader;
 
 @Component
 @Slf4j
 public class JarvisRunnerService {
     
     @Autowired
-    private JarvisConfigLoader configLoader;
+    private ConfigLoader configLoader;
     
     @Autowired
     private InitActionService initActionService;
