+++ src/main/java/ai/aliz/jarvis/service/assertaction/AssertActionService.java
@@ -15,7 +15,7 @@ import ai.aliz.jarvis.context.JarvisContextLoader;
 import ai.aliz.jarvis.context.JarvisContext;
 import ai.aliz.jarvis.context.JarvisContextType;
 
-import ai.aliz.jarvis.jarvisconfig.AssertActionConfig;
+import ai.aliz.jarvis.config.AssertActionConfig;
 
 @Component
 @Slf4j
