+++ src/main/java/ai/aliz/jarvis/service/shared/platform/SFTPService.java
@@ -22,7 +22,7 @@ import com.jcraft.jsch.SftpException;
 import org.springframework.stereotype.Service;
 
 import ai.aliz.jarvis.context.JarvisContext;
-import ai.aliz.jarvis.jarvisconfig.InitActionConfig;
+import ai.aliz.jarvis.config.InitActionConfig;
 
 import static ai.aliz.jarvis.util.JarvisConstants.HOST;
 import static ai.aliz.jarvis.util.JarvisConstants.PASSWORD;
