+++ src/test/java/ai/aliz/jarvis/integration/TestInitActionServiceBQIntegration.java
@@ -25,8 +25,8 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import ai.aliz.jarvis.context.JarvisContextLoader;
 import ai.aliz.jarvis.service.initaction.InitActionService;
-import ai.aliz.jarvis.jarvisconfig.InitActionConfig;
-import ai.aliz.jarvis.jarvisconfig.InitActionConfigFactory;
+import ai.aliz.jarvis.config.InitActionConfig;
+import ai.aliz.jarvis.config.InitActionConfigFactory;
 
 import org.junit.Before;
 import org.junit.Rule;
